release: v0.14 — ml-dsa 0.1.0 + full audit remediation + v0.13->v0.14 epoch
Workflow-driven audits (bug / profiler-optimization / security) surfaced
11 + 21 + 2 confirmed findings; this lands the remediation, the ml-dsa
stabilization, and the resulting chain-epoch version advance as one
release. v0.13-epoch WALs do not replay under v0.14 (new chain epoch).

Dependencies
- ml-dsa =0.1.0-rc.9 -> =0.1.0 (NIST FIPS 204 final). KeyGen trait removed
  -> SigningKey::<MlDsa65>::from_seed; signature/key bytes unchanged.
- zeroize feature enabled on ml-dsa + ed25519-dalek; zeroize added as a
  direct dep for in-kernel seed scrub.

Version epoch (all four axes -> 0.14)
- package versions, WalHeader::CURRENT_KERNEL_SEMVER (0,14,0),
  ABI_VERSION (0,14), DOMAIN_CTX chain-separation literal "v0.14".

Security
- Authenticated WAL verification: TrustAnchor + verify_chain_anchored /
  replay_into_verified reject signature downgrade, verifying-key
  substitution, manifest mismatch, and tail truncation (closes the
  critical header-trust gap; caller owns the trust-root policy).
- Ed25519 verify_strict (malleability); WAL-record signature domain
  separation (WAL_SIG_DOMAIN || body); snapshot deserialize_verified +
  structural validation on decode; replay A14 postcard/blake3 gates;
  quota enforcement (max_entities / max_scheduled + submit back-pressure).

Correctness
- ResourceLedger rebuilt on component_sizes as the single accounting
  authority: replace delta (no double-count), despawn cascade, saturating,
  stored-size remove.
- Budget projection u64->i64 overflow hardening (oversized size cannot
  wrap past the gate); scheduler saturating + three-table index validate;
  replay caps from_bits_retain; WAL actor field chain-hashed; ObserverHandle
  widened to u64.

Consistency
- Layer A catastrophic-invariant count reconciled 8 -> 7 across docs;
  supply-chain/audits.toml orphan crates (aes-gcm, chacha20poly1305)
  dropped + ml-dsa audit note updated; deny.toml dual RustCrypto
  trait-stack documented.

Layer A escalations recorded in formal/axiom-test-cite.toml
(item_1 + item_7 escalation_event_2). Baselines regenerated: 4 frozen-hex,
ci/l0-baseline-hashes.txt, ci/test-baselines.txt (227 -> 241).

Verified locally: cargo test 241/0-fail, clippy --all-targets clean,
verify-l0-baseline OK, verify-axiom-cite OK, dice D1-Total replay OK,
adversarial fix-verification 11/11 sound.

Downstream: the sibling ArkheForge hybrid_and_mode Kani harness should
re-confirm against v0.14 (signed-message construction changed; the
AND-mode property is preserved by construction).

+++ arkhe-kernel/src/runtime/event.rs
@@ -130,8 +130,11 @@ pub enum DeferReason {
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
 #[serde(transparent)]
 pub struct ObserverHandle(
-    /// Monotonic registry index assigned at registration.
-    pub u16,
+    /// Monotonic registry index assigned at registration. A `u64` (not
+    /// `u16`) so the monotonic counter cannot realistically saturate and
+    /// alias a live handle (the registry has no unregister, so handles
+    /// only ever grow).
+    pub u64,
 );
 
 bitflags! {
